requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Login==0.6.3
Authlib==1.2.1
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1